Use new index and database library

Closes #1971, #2444, #2443, #2442, #2436, #2412, #2370, #2353, #2322, #2157, #2081, #2080, #1989, #1710, #1588, #1206, #860, and #649

See merge request fdroid/fdroidclient!1135

+++ app/src/main/java/org/fdroid/fdroid/installer/InstallHistoryService.java
@@ -112,7 +112,7 @@ public class InstallHistoryService extends IntentService {
         long timestamp = System.currentTimeMillis();
         Apk apk = intent.getParcelableExtra(Installer.EXTRA_APK);
         String packageName = apk.packageName;
-        int versionCode = apk.versionCode;
+        long versionCode = apk.versionCode;
 
         List<String> values = new ArrayList<>(4);
         values.add(String.valueOf(timestamp));
